separate files in jwt-service
