+++ internal/service/jwt-service/init.go
@@ -1,19 +1,8 @@
 package jwtservice
 
 import (
-	"errors"
-	"github.com/golang-jwt/jwt/v5"
 	"github.com/sladkoezhkovo/auth-service/internal/configs"
-	"github.com/sladkoezhkovo/auth-service/internal/entity"
-	"os"
-	"time"
 )
-
-type userClaims struct {
-	Role  int
-	Email string
-	jwt.RegisteredClaims
-}
 
 type TokenStorage interface {
 	Get(key string) (string, error)
@@ -31,82 +20,6 @@ func New(config *configs.JwtConfig, storage TokenStorage) *jwtService {
 		config:  config,
 		storage: storage,
 	}
-}
-
-func (s *jwtService) generateAccess(user *entity.User) (string, error) {
-	claims := userClaims{
-		Role:  user.Role,
-		Email: user.Email,
-	}
-	claims.ExpiresAt = jwt.NewNumericDate(time.Now().Add(time.Hour * time.Duration(s.config.AccessTTL)))
-	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
-	signed, err := token.SignedString(os.Getenv("JWT_ACCESS_SECRET"))
-	if err != nil {
-		return "", err
-	}
-
-	return signed, nil
-}
-
-func (s *jwtService) generateRefresh(user *entity.User) (string, error) {
-	claims := userClaims{
-		Role:  user.Role,
-		Email: user.Email,
-	}
-	claims.ExpiresAt = jwt.NewNumericDate(time.Now().Add(time.Hour * time.Duration(s.config.RefreshTTL)))
-	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
-	signed, err := token.SignedString(os.Getenv("JWT_ACCESS_SECRET"))
-	if err != nil {
-		return "", err
-	}
-
-	return signed, nil
-}
-
-func (s *jwtService) Generate(user *entity.User) (*entity.Tokens, error) {
-
-	refresh, err := s.generateRefresh(user)
-	if err != nil {
-		return nil, err
-	}
-
-	access, err := s.generateAccess(user)
-	if err != nil {
-		return nil, err
-	}
-
-	return &entity.Tokens{
-		Access:  access,
-		Refresh: refresh,
-	}, nil
-}
-
-func (s *jwtService) validate(tokenString string, secret string) (*entity.UserClaims, error) {
-
-	token, err := jwt.ParseWithClaims(tokenString, &userClaims{}, func(token *jwt.Token) (interface{}, error) {
-		return secret, nil
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	claims, ok := token.Claims.(userClaims)
-	if !ok {
-		return nil, errors.New("unable to parse claims")
-	}
-
-	return &entity.UserClaims{
-		Email: claims.Email,
-		Role:  claims.Role,
-	}, nil
-}
-
-func (s *jwtService) ValidateRefresh(tokenString string) (*entity.UserClaims, error) {
-	return s.validate(tokenString, os.Getenv("JWT_REFRESH_SECRET"))
-}
-
-func (s *jwtService) ValidateAccess(tokenString string) (*entity.UserClaims, error) {
-	return s.validate(tokenString, os.Getenv("JWT_ACCESS_SECRET"))
 }
 
 func (s *jwtService) Save(email, token string) error {
